+++ test/libsolidity/smtCheckerTests/loops/for_loop_trivial_condition_3.sol
@@ -15,4 +15,3 @@ contract C {
 }
 // ----
 // Warning: (115-121): Unused local variable.
-// Warning: (356-370): Assertion violation happens here